src/main/java/com/cs320/milestones/Task.java
package com.cs320.milestones;
import lombok.Data;
import lombok.NoArgsConstructor;

//@Data
//@NoArgsConstructor
// set Args through lombok
public class Task {
    private String taskID;
    private String name;
    private String description;

    public String getTaskID() {
        return taskID;
    }

    public void setTaskID(String taskID) {
        this.taskID = taskID;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getDescription() {
        return description;
    }

    public void setDescription(String description) {
        this.description = description;
    }

    public Task(String taskID, String name, String description) {
        if (taskID == null || taskID.length() > 10) {
            throw new IllegalArgumentException("Invalid task ID");
        }
        if (name == null || name.length() > 20) {
            throw new IllegalArgumentException("Invalid name");
        }
        if (description == null || description.length() > 50) {
            throw new IllegalArgumentException("Invalid description");
        }

        this.taskID = taskID;
        this.name = name;
        this.description = description;
    }
}
